requirements.txt
streamlit
openai
sqlmodel
pydantic
apscheduler
akshare
playwright
python-dotenv
pandas
beautifulsoup4
httpx
feedparser
